Add Docker support and update project settings
Updated `BookStore.Api.csproj` to include `UserSecretsId` and `DockerDefaultTargetOS` properties, and added a package reference for `Microsoft.VisualStudio.Azure.Containers.Tools.Targets`. Reformatted `launchSettings.json` and added a Docker profile with environment variables for HTTPS and HTTP ports. Created a `.dockerignore` file to specify files and directories to be ignored by Docker. Added a `Dockerfile` with multiple stages for building and running the application in a Docker container.